Adds more generic item-allocation queries
- Creates the `omicron_nexus::db::queries` submodule, for custom /
  complex queries, organized around specific resources
- Adds a general-purpose `NextItem` query, which looks for the first
  available item in a database column, with a defined starting point and
  maximum offset / scan size. This can be used to implement a linear
  search over a defined space. One of the benefits is that for
  randomly-selected items, we can avoid a retry loop, by instead
  starting from the random base item and inserting the first successful
  value. There are several places where we use a similar query, or use a
  retry loop that could be replaced by this query.
- Use the `NextGuestIpv4Address` query when creating NICs.
- Use the `NextNicSlot` query when creating guest NICs
